Massively simplify this
There's no need for most of what this was doing; the functionality already existed, the goal was simply to put it somewhere else on the page and then style it. In particular, making N+1 requests (where N is the number of answers on the page) is a recipe for getting mods IP-banned; that info is already being loaded, just use it!